Fix distance being float64 with numpy v2

+++ src/silx/gui/plot/backends/glutils/GLPlotCurve.py
@@ -546,7 +546,8 @@ def distancesFromArrays(xData, yData, ratio: float = 1.0):
                 (
                     numpy.ediff1d(xData[begin:end], to_begin=numpy.float32(0.0)),
                     numpy.ediff1d(
-                        yData[begin:end] * ratio, to_begin=numpy.float32(0.0)
+                        (yData[begin:end] * ratio).astype(numpy.float32),
+                        to_begin=numpy.float32(0.0)
                     ),
                 )
             )[0]
